requirements.txt
leancloud-sdk
Flask==0.10.1
Flask-RESTful==0.3.2
Flask-wechatpy==0.1.3
wechatpy==1.5.0
Jinja2==2.7.3
wheel==0.24.0
cryptography>=0.8.2
gevent>=1.0.2
Werkzeug>=0.11.11
gevent-websocket>=0.9.5
